Day 2 Part 2: Coroutines search!
OK this works OK and is seems fast at 0.098s. Maybe we just find the
result early in our search and maybe 10,000 isn't that big of a search
space for the problem. Unfortunately IntelliJ is showing the coroutine
stuff as red because it thinks I'm using Kotlin 1.2. The program compiles
in the command line and runs but I gotta fix the IDE display.